+++ app/src/main/java/com/ftw/hometerview/di/ui/MainFragmentViewModelModule.kt
@@ -4,15 +4,12 @@ import com.ftw.domain.usecase.favorite.GetFavoriteBuildingsUseCase
 import com.ftw.domain.usecase.favorite.GetFavoriteReviewsUseCase
 import com.ftw.domain.usecase.review.GetLocationReviewsUseCase
 import com.ftw.domain.usecase.user.GetCachedUserUseCase
-import com.ftw.domain.usecase.writtenreview.GetWrittenReviewsUseCase
 import com.ftw.hometerview.dispatcher.Dispatcher
 import com.ftw.hometerview.ui.main.favorite.favoritelist.FavoriteBuildingsViewModel
 import com.ftw.hometerview.ui.main.favorite.favoritelist.FavoriteReviewsViewModel
 import com.ftw.hometerview.ui.main.home.HomeViewModel
 import com.ftw.hometerview.ui.main.home.review.LocationReviewListViewModel
 import com.ftw.hometerview.ui.main.mypage.MyPageViewModel
-import com.ftw.hometerview.ui.writtenreview.nonreview.NonReviewViewModel
-import com.ftw.hometerview.ui.writtenreview.writtenreviewlist.WrittenReviewListViewModel
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -88,21 +85,4 @@ class MainFragmentViewModelModule {
         )
     }
 
-    @Provides
-    @FragmentScoped
-    fun provideWrittenReviewListViewModel(
-        dispatcher: Dispatcher,
-        getWrittenReviewsUseCase: GetWrittenReviewsUseCase
-    ): WrittenReviewListViewModel {
-        return WrittenReviewListViewModel(
-            dispatcher,
-            getWrittenReviewsUseCase
-        )
-    }
-
-    @Provides
-    @FragmentScoped
-    fun provideNonReviewViewModel(): NonReviewViewModel {
-        return NonReviewViewModel()
-    }
 }
