+++ app/src/main/res/layout/activity_written_reviews.xml
@@ -8,8 +8,9 @@
 
         <variable
             name="viewModel"
-            type="com.ftw.hometerview.ui.writtenreview.WrittenReviewViewModel"
+            type="com.ftw.hometerview.ui.writtenreview.WrittenReviewsViewModel"
             />
+        <import type="android.view.View"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -22,8 +23,8 @@
             android:id="@+id/back_button"
             android:layout_width="@dimen/dp_size_32"
             android:layout_height="@dimen/dp_size_32"
-            android:layout_marginTop="@dimen/dp_size_16"
             android:layout_marginStart="@dimen/dp_size_16"
+            android:layout_marginTop="@dimen/dp_size_16"
             android:src="@drawable/ic_baseline_navigate_before_24"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
@@ -54,61 +55,69 @@
             app:layout_constraintTop_toBottomOf="@+id/written_hometerview_text"
             />
 
-        <FrameLayout
-            android:id="@+id/fragment_view"
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/my_review_view"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@+id/end_view"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/divide_view"
-            />
-<!--        <androidx.recyclerview.widget.RecyclerView-->
-<!--            android:id="@+id/written_review_recyclerview"-->
-<!--            android:layout_width="match_parent"-->
-<!--            android:layout_height="0dp"-->
-<!--            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"-->
-<!--            app:layout_constraintBottom_toTopOf="@+id/end_view"-->
-<!--            app:layout_constraintEnd_toEndOf="parent"-->
-<!--            app:layout_constraintStart_toStartOf="parent"-->
-<!--            app:layout_constraintTop_toBottomOf="@+id/divide_view"-->
-<!--            />-->
+            android:visibility="@{viewModel.writtenRevieewItems.empty ? View.GONE : View.VISIBLE}"
+            >
 
-<!--        <androidx.constraintlayout.widget.ConstraintLayout-->
-<!--            android:layout_width="match_parent"-->
-<!--            android:layout_height="0dp"-->
-<!--            app:layout_constraintBottom_toTopOf="@+id/end_view"-->
-<!--            app:layout_constraintEnd_toEndOf="parent"-->
-<!--            app:layout_constraintStart_toStartOf="parent"-->
-<!--            app:layout_constraintTop_toBottomOf="@+id/divide_view"-->
-<!--            >-->
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/written_review_recyclerview"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                app:layout_constraintBottom_toTopOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="parent"
+                />
+        </androidx.constraintlayout.widget.ConstraintLayout>
 
-<!--            <TextView-->
-<!--                android:id="@+id/non_review_text"-->
-<!--                android:layout_width="wrap_content"-->
-<!--                android:layout_height="wrap_content"-->
-<!--                android:fontFamily="@font/pretendard_regular"-->
-<!--                android:textColor="@color/gray_600"-->
-<!--                app:layout_constraintEnd_toEndOf="parent"-->
-<!--                app:layout_constraintStart_toStartOf="parent"-->
-<!--                app:layout_constraintTop_toTopOf="parent"-->
-<!--                tools:text="작성한 집터뷰가 없어요"-->
-<!--                />                      -->
 
-<!--&lt;!&ndash;            <com.ftw.hometerview.design.MainButton&ndash;&gt;-->
-<!--&lt;!&ndash;                android:id="@+id/write_review_button"&ndash;&gt;-->
-<!--&lt;!&ndash;                android:layout_width="wrap_content"&ndash;&gt;-->
-<!--&lt;!&ndash;                android:layout_height="wrap_content"&ndash;&gt;-->
-<!--&lt;!&ndash;                android:fontFamily="@font/pretendard_regular"&ndash;&gt;-->
-<!--&lt;!&ndash;                android:textColor="@color/gray_600"&ndash;&gt;-->
-<!--&lt;!&ndash;                app:layout_constraintBottom_toottomOf="parent"&ndash;&gt;-->
-<!--&lt;!&ndash;                app:layout_constraintEnd_toEndOf="parent"&ndash;&gt;-->
-<!--&lt;!&ndash;                app:layout_constraintStart_toStartOf="parent"&ndash;&gt;-->
-<!--&lt;!&ndash;                app:layout_constraintTop_toBottomOf="@+id/non_review_text"&ndash;&gt;-->
-<!--&lt;!&ndash;                android:text="집터뷰 작성하기"&ndash;&gt;-->
-<!--&lt;!&ndash;                />&ndash;&gt;-->
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/non_my_review_view"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintBottom_toTopOf="@+id/end_view"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/divide_view"
+            android:visibility="@{viewModel.writtenRevieewItems.empty ? View.VISIBLE : View.GONE}"
+            >
 
-<!--        </androidx.constraintlayout.widget.ConstraintLayout>-->
+            <TextView
+                android:id="@+id/non_review_text"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:fontFamily="@font/pretendard_regular"
+                android:text="작성한 집터뷰가 없어요"
+                android:textColor="@color/gray_600"
+                android:layout_marginBottom="@dimen/dp_size_60"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                tools:text="작성한 집터뷰가 없어요"
+                />
+
+            <com.ftw.hometerview.design.MainButton
+                android:id="@+id/write_review_button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="집터뷰 작성하기"
+                android:onClick="@{() -> viewModel.onClickWriteReview()}"
+                android:layout_marginTop="@dimen/dp_size_16"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/non_review_text"
+                />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
 
         <View
             android:id="@+id/end_view"
