--- app/src/main/res/layout/fragment_written_review_list.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    >
-
-    <data>
-        <variable
-            name="viewModel"
-            type="com.ftw.hometerview.ui.writtenreview.writtenreviewlist.WrittenReviewListViewModel"
-            />
-    </data>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        tools:listitem="@layout/list_item_favorite_review"
-        tools:itemCount="10"
-        />
-</layout>
